warning as error (#269)
* warnings as errors
disable warnings for 3rd party dependencies

* fix more warnings

* linux callback

* shuffle includes

* different version to check for msvc

* unsigned in loop

* clang 9 linux
unsigned loop

* mac/ios

* back to clang 8 for Linux

* missing bgfx include

* arcana update
test with fork of base-n

* c99 extensions

* update basen from own repo

* couple more warnings...

* unused mac

* more warnings

* linux gcc

* more linux stuff

* disable warning 5205 for msvc

* catch by value
disable 2 warnings for clang

* typo

* update base-n
fix unused constexpr in nativexr

* constexpr nativexr

* arcana again for gcc

* disable link warning for bx

* target_link_options

* link option

* update bgfx

* dummy

* update bgfx.cmake

* upgrade bgfx

* dummy

* link flag again

* link flag

* flag...

* update bgfx.cmake

* once more

* revert flag change in bgfx.cmake

* bgfx.cmake update

* PR comments

* PR comments

* warnings xr

* struct/class mismatch

* more mismatch

* static_cast

* win32 warning

* static_cast

* base-n PR merged

* base-n update

* more casts

+++ Plugins/NativeEngine/Source/NativeEngineAPI.cpp
@@ -7,7 +7,7 @@ namespace Babylon::Plugins::NativeEngine
 {
     void InitializeGraphics(void* windowPtr, size_t width, size_t height)
     {
-        Babylon::NativeEngine::InitializeWindow(windowPtr, width, height);
+        Babylon::NativeEngine::InitializeWindow(windowPtr, static_cast<uint32_t>(width), static_cast<uint32_t>(height));
     }
 
     void Initialize(Napi::Env env)
@@ -24,7 +24,7 @@ namespace Babylon::Plugins::NativeEngine
         pd.backBuffer = nullptr;
         pd.backBufferDS = nullptr;
         bgfx::setPlatformData(pd);
-        bgfx::reset(width, height);
+        bgfx::reset(static_cast<uint32_t>(width), static_cast<uint32_t>(height));
 
         auto& window = Plugins::Internal::NativeWindow::GetFromJavaScript(env);
         window.Resize(width, height);
